closer to merge completion

+++ tests/testfilesetup.py
@@ -2,7 +2,7 @@ import sys
 import os
 import json
 import pickle
-from physcraper import wrappers, generate_ATT_from_files, AlignTreeTax, ConfigObj, IdDicts, PhyscraperScrape
+from physcraper import OtuJsonDict, generate_ATT_from_files, AlignTreeTax, ConfigObj, IdDicts, PhyscraperScrape
 #
 
 
@@ -11,7 +11,7 @@ mattype="fasta"
 trfn= "tests/data/tiny_test_example/test.tre"
 schema_trf = "newick"
 workdir="tests/output/owndata"
-configfi = "tests/data/localblast.config"
+configfi = "tests/data/test.config"
 id_to_spn = r"tests/data/tiny_test_example/test_nicespl.csv"
 otu_jsonfi = "{}/otu_dict.json".format(workdir)
 
@@ -21,10 +21,13 @@ otu_jsonfi = "{}/otu_dict.json".format(workdir)
 if not os.path.exists("{}".format(workdir)):
 	os.makedirs("{}".format(workdir))
 
+conf = ConfigObj(configfi)
+ids = IdDicts(conf, workdir=workdir)
 
-otu_json = wrappers.OtuJsonDict(id_to_spn, configfi)
-with open(otu_jsonfi,"w") as outfile:
-    json.dump(otu_json, outfile)
+if not os.path.exists(otu_jsonfi):
+    otu_json = OtuJsonDict(id_to_spn, ids)
+    with open(otu_jsonfi,"w") as outfile:
+        json.dump(otu_json, outfile)
 
 
 
@@ -38,9 +41,8 @@ data_obj = generate_ATT_from_files(seqaln=seqaln,
 
 data_obj.prune_short()
 data_obj.dump(filename = "tests/data/tiny_dataobj.p")
-conf = ConfigObj(configfi)
-ids = IdDicts(conf, workdir=workdir)
-scraper =  PhyscraperScrape(data_obj, ids, conf)
+
+scraper =  PhyscraperScrape(data_obj, ids)
 scraper.read_blast(blast_dir="tests/data/tiny_test_example/blast_files")
 scraper.remove_identical_seqs()
